+++ lib/sstls/archive.go
@@ -5,7 +5,7 @@ package sstls
  * Read and Save certs with an archive file
  * By J. Stuart McMurray
  * Created 20240327
- * Last Modified 20260817
+ * Last Modified 20260823
  */
 
 import (
@@ -38,12 +38,17 @@ func LoadCachedCertificate(certFile string) (tls.Certificate, error) {
 			err,
 		)
 	}
-	return loadCachedCertificate(certFile, b)
+	return ParseCachedCertificate(b)
 }
 
-// loadCachedCertificate does what LoadCachedCertificate says it does but
-// from a pre-loaded file.  certFile is only used in error messages.
-func loadCachedCertificate(certFile string, b []byte) (tls.Certificate, error) {
+// ParseCachedCertificate is like [LoadCachedCertificate], but loads the
+// certificate from b, which should have been created with [SaveCertificate].
+func ParseCachedCertificate(b []byte) (tls.Certificate, error) {
+	/* Mooched from the go stdlib .*/
+	fail := func(err error) (tls.Certificate, error) {
+		return tls.Certificate{}, err
+	}
+
 	/* Read the saved cert. */
 	ta := txtar.Parse(b)
 
@@ -60,26 +65,20 @@ func loadCachedCertificate(certFile string, b []byte) (tls.Certificate, error) {
 
 	/* Try to use it. */
 	if 0 == len(certB) {
-		return tls.Certificate{}, ErrCacheFileEmpty
+		return fail(ErrCacheFileEmpty)
 	} else if 0 == len(keyB) {
-		return tls.Certificate{}, fmt.Errorf(
-			"PEM-encoded key missing",
-		)
+		return fail(ErrPrivateKeyPEMEmpty)
 	}
 	cert, err := tls.X509KeyPair(certB, keyB)
 	if nil != err {
-		return tls.Certificate{}, fmt.Errorf(
-			"loading certificate from %s: %w",
-			certFile,
-			err,
-		)
+		return fail(fmt.Errorf("parsing certificate: %w", err))
 	}
 
 	/* Make sure Leaf is set.  At one point, tls.X509KeyPair didn't
 	always do this, but it should now unless someone set
 	GODEBUG=x509keypairleaf=0. */
 	if nil == cert.Leaf {
-		return tls.Certificate{}, ErrLeafCertificateNotSet
+		return fail(ErrLeafCertificateNotSet)
 	}
 
 	return cert, nil