+++ lib/sstls/archive_test.go
@@ -5,11 +5,12 @@ package sstls
  * Read and Save certs with an archive file
  * By J. Stuart McMurray
  * Created 20240327
- * Last Modified 20260111
+ * Last Modified 20260823
  */
 
 import (
 	"crypto/x509"
+	"errors"
 	"os"
 	"path/filepath"
 	"testing"
@@ -172,4 +173,41 @@ func TestSaveCertificate_TruncateFile(t *testing.T) {
 			want,
 		)
 	}
+}
+
+// Do we get the right errors parsing invalid certificates?
+func TestParseCachedCertificate_Errors(t *testing.T) {
+	for n, want := range map[string]error{
+		"empty_cert": ErrCacheFileEmpty,
+		"empty_key":  ErrPrivateKeyPEMEmpty,
+	} {
+		t.Run(n, func(t *testing.T) {
+			if _, got := ParseCachedCertificate(
+				[]byte(mustTDFile(t)),
+			); !errors.Is(got, want) {
+				t.Errorf(
+					"Incorrect parse error\n"+
+						" got: %v\n"+
+						"want: %v",
+					got,
+					want,
+				)
+			}
+		})
+	}
+}
+
+// Do we get the right sort of error if we can't parse the cert's PEM?
+func TestParseCcahedCertificate_ParseError(t *testing.T) {
+	_, err := ParseCachedCertificate([]byte(mustTDFile(t)))
+	/* tls library gives us a string :( */
+	if got, want := err.Error(), "parsing certificate: tls: failed to "+
+		"find any PEM data in certificate input"; got != want {
+		t.Errorf(
+			"Incorrect parse error\n got: %v\nwant: %v",
+			got,
+			want,
+		)
+	}
+
 }